\name{lineRatio-package}
\alias{lineRatio-package}
\alias{lineRatio}
\docType{package}
\title{
Provides function to calculate a line as a set of points that is the results the ratio other two lines.
}
\description{
Provides function to calculate a line as a set of points that is the results the ratio other two lines.
}
\details{
\tabular{ll}{
Package: \tab lineRatio\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-20\cr
License: \tab GPLv3\cr
}
}
\author{Valentin Ruano-Rubio <valentin.ruano@gmail.com>}
\references{}
\keyword{ ratio }
\seealso{
}
\examples{
}
